requirements.txt
# ============================================================
# Core — Object Detection
# ============================================================
ultralytics>=8.3.0          # YOLOv8/v9/v10/v11 통합 패키지 (AGPL-3.0)

# ============================================================
# Deep Learning Framework
# ============================================================
# torch/torchvision: GPU 환경은 conda 또는 별도 CUDA 빌드 권장
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.2.0
torchvision>=0.17.0

# ============================================================
# Computer Vision
# ============================================================
opencv-python>=4.9.0        # 이미지 I/O, 전처리, 시각화
Pillow>=10.2.0               # 보조 이미지 처리

# ============================================================
# Data & Analysis
# ============================================================
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# ============================================================
# Visualization
# ============================================================
matplotlib>=3.8.0
seaborn>=0.13.0

# ============================================================
# Experiment Tracking (선택 — 팀 논의 후 활성화)
# ============================================================
# wandb>=0.16.0

# ============================================================
# Configuration
# ============================================================
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ============================================================
# Code Quality
# ============================================================
ruff>=0.4.0
black>=24.3.0

# ============================================================
# Test
# ============================================================
pytest>=8.0.0

# ============================================================
# Notebook
# ============================================================
ipykernel>=6.29.0
tqdm>=4.66.0
